updated app.js, created index.html and page template.js files
